+++ CHANGELOG.md
@@ -37,6 +37,9 @@ nil.
 - Add Address#concatenated_address1 and Address#concatenated_address2 methods [#158](https://github.com/Shopify/worldwide/pull/158)
 - Updated root translation keys for regional locales `bg-BG`, `hr-HR`, `lt-LT`, `ro-RO`, `sk-SK`, `sl-SI`  to the base locale (e.g. `bg`) [#164](https://github.com/Shopify/worldwide/pull/164).
 - Remove city from UAE address form and enable city autofill [#159](https://github.com/Shopify/worldwide/pull/159)
+- Support additional fields in IL, update streetNumber and neighborhood requirement for BE, CL, ES, MX [#166](https://github.com/Shopify/worldwide/pull/166)
+
+---
 
 ## [0.12.2] - 2024-05-21