+++ CHANGELOG.md
@@ -27,9 +27,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+---
+
+## [0.14.0] - 2024-05-29
+
 - Add support for a line2 address field [#173](https://github.com/Shopify/worldwide/pull/173)
 - Remove city from UAE address form and enable city autofill [#175](https://github.com/Shopify/worldwide/pull/175)
----
 
 ## [0.13.0] - 2024-05-27